refactor(debug): rename formating.py to formatting.py with backwards-compat shim
The module was previously misspelled as ``formating``. The canonical file
is now :mod:`graphene_django.debug.exception.formatting`, and
``middleware.py`` imports from the corrected path.

A thin shim is left in place at the legacy ``formating`` import path
re-exporting :func:`wrap_exception` so any third-party code importing
the misspelled module continues to work. The shim is documented as
deprecated.

Made-with: Cursor
